+++ roundcubeWebmail/utils_paths.py
@@ -1,34 +1,32 @@
 # -*- coding: utf-8 -*-
-"""Shared paths and helpers for Roundcube Webmail plugin."""
+"""Shared paths for Roundcube Webmail catalog entry (host-package redirect)."""
 from __future__ import print_function
 
 import os
 
 PLUGIN_NAME = 'roundcubeWebmail'
-SETTINGS_FILE = '/home/cyberpanel/plugins/roundcubeWebmail_settings.json'
-DISABLE_MARKER = '/usr/local/CyberCP/public/roundcube/.cp_webmail_disabled'
-
-ROUNDCUBE_ROOT = '/usr/local/CyberCP/public/roundcube'
+# CPN host package paths (never CyberPanel).
+ROUNDCUBE_ROOT = '/opt/cpn-webmail/roundcube'
 ROUNDCUBE_PUBLIC = os.path.join(ROUNDCUBE_ROOT, 'public_html')
 ROUNDCUBE_CONFIG = os.path.join(ROUNDCUBE_ROOT, 'config', 'config.inc.php')
-ROUNDCUBE_DB_DIR = os.path.join(ROUNDCUBE_ROOT, 'db')
-ROUNDCUBE_SQLITE = os.path.join(ROUNDCUBE_DB_DIR, 'roundcube.db')
-ROUNDCUBE_VERSION_FILE = '/etc/cyberpanel/roundcube_version'
-ROUNDCUBE_FALLBACK_VERSION = '1.6.9'
+ROUNDCUBE_DB_DIR = ROUNDCUBE_ROOT
+ROUNDCUBE_SQLITE = os.path.join(ROUNDCUBE_ROOT, 'db.sqlite')
+SETTINGS_FILE = '/var/lib/cpn/plugins/roundcubeWebmail_settings.json'
+DISABLE_MARKER = '/opt/cpn-webmail/roundcube/.cp_webmail_disabled'
+ROUNDCUBE_VERSION_FILE = '/var/lib/cpn/roundcube_version'
+ROUNDCUBE_FALLBACK_VERSION = '1.7.3'
 
 LSWS_ROOT = '/usr/local/lsws'
-VHOST_DIR = os.path.join(LSWS_ROOT, 'conf', 'vhosts', 'CyberPanel')
-VHOST_CONF = os.path.join(VHOST_DIR, 'vhost.conf')
-BIND_CONF = '/usr/local/lscp/conf/bind.conf'
+VHOST_DIR = os.path.join(LSWS_ROOT, 'conf', 'vhosts', 'CPNWebmail')
+VHOST_CONF = os.path.join(VHOST_DIR, 'vhconf.conf')
+BIND_CONF = '/var/lib/cpn/bind.conf'
 
 
 def _log(msg):
     try:
-        from plogical.CyberCPLogFileWriter import CyberCPLogFileWriter
-        CyberCPLogFileWriter.writeToFile('[roundcubeWebmail] ' + str(msg))
+        print('[roundcubeWebmail] ' + str(msg))
     except Exception:
         pass
-    print('[roundcubeWebmail] ' + str(msg))
 
 
 def detect_lsphp_version():
@@ -49,4 +47,4 @@ def get_panel_port():
                     return port
     except Exception:
         pass
-    return '8090'
+    return '2087'